Refactor utils file and remove unnecessary functions
Change-type: patch
