+++ src/components/DataTypes/array.tsx
@@ -1,7 +1,7 @@
 import { JSONSchema6 } from 'json-schema';
 import assign from 'lodash/assign';
 import * as React from 'react';
-import * as utils from '../../utils';
+import { randomString } from '../../utils';
 import { DataTypeEditProps } from '../Filters';
 import Input, { InputProps } from '../Input';
 import Textarea, { TextareaProps } from '../Textarea';
@@ -107,7 +107,7 @@ export const createFilter = (
 ): JSONSchema6 => {
 	const { title } = schema;
 	const base: StringFilter = {
-		$id: utils.randomString(),
+		$id: randomString(),
 		title: operator,
 		description: getJsonDescription(
 			title || field,
